defender_api/requirements.txt
# gcp
functions-framework==3.*

# google cloud
google-cloud
google-cloud-firestore
google-cloud-secret-manager

# packages
gspread
phonenumbers
whatsapp_api_client_python